+++ app/views/topics/show.html.erb
@@ -1,5 +1,5 @@
 <% content_for(:title, "#{@topic.title}") %>
-<% set_meta_tags(og: {image: ogp_image_topic_path(@topic , host: request.base_url)})  %>
+<% set_meta_tags(twitter: {image: ogp_image_topic_url(@topic)})  %>
 <div class="px-4 py-10">
   <div class="bg-white shadow-lg rounded-2xl w-full py-8 px-4 space-y-6">
     <div class="text-center mb-5">
